[FABG-913] Fail-fast for Discovery Service
When a request is sent to the Discovery Service and MaxTargets is set to > 1
then, even though a response is received from one peer, we wait for ALL other
requests to complete. When a peer is down then (depending on the connection
settings) this may take a long time given GRPC's retry logic.

This patch changes the logic to cancel all other outstanding requests as soon
as a successful response is received.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>
Change-Id: I2c2f22b66aca85995a1f007b113aca10c237d3eb