[FABG-916] Evict connection in TRANSIENT_FAILURE state
When a peer is shut down, the connection to that peer goes into TRANSIENT_FAILURE
state and we wait for it to be set to READY state. After the peer is started the
connection will be in TRANSIENT_FAILURE state for several more seconds (even though
the peer is up), potentially causing timeouts at the client.

With this patch, connections that are in TRANSIENT_FAIURE state are evicted from
the cache. An error is immediately returned to the client so that it may retry
connecting with a fresh connection.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>
Change-Id: I4936fd2fc219f04761e035345b2914461b76a361